Implemented lobby transition between the two separate views.
Cleaned up code slightly. The amount of lines needed to position each
component is a bit unsightly and ugly, but the GridBag layout gets us
the closest to what the screenshot positioning looked like so I'm
keeping it for now. If it's too much to handle, we can down grade to a
less complicated layout.