Added CardPanel class that houses the cards and its contents.
There is a test main method attached to this class as well. Running it
will render something fairly ugly, but it doesn't have bearing on the
actual class because it's just generating a temporary frame to house the
card. Otherwise, there are a few listeners and event firing methods that
might be useful to look at. They're incomplete, but they're meant to add
a border to the cards when they're selectable and when they're selected.

Need to add the pictures to a common directory and fix pathing to the
images.