Cleanup, Resizing, and integrated the HandPanel.
Cleaned up some unused code and added some test methods so that we don't
have to deal with spawning multiple clients each time we test. Added the
HandPanel which houses the players cards - It needs to hold a max of 6
cards (3 player game). Did some resizing in the MainPanel. We need to
settle on how bit we want the client to be so that we can work of those
dimensions. It's not feasible to try and get things to resize
dynamically.